// src/common/config/data/enum/config-key.enum.ts
export enum ConfigKey {
  NodeEnv = 'NODE_ENV',
  AppName = 'APP_NAME',
  AppPort = 'APP_PORT',
  AppBaseUrl = 'APP_BASE_URL',
  AppCorsOrigin = 'APP_CORS_ORIGIN',
  AppTrustProxy = 'APP_TRUST_PROXY',
  AppHttpPayloadErrorCode = 'APP_HTTP_PAYLOAD_ERROR_CODE',
  LogLevel = 'LOG_LEVEL',
  SwaggerEnabled = 'SWAGGER_ENABLED',
  SwaggerTitle = 'SWAGGER_TITLE',
  SwaggerDescription = 'SWAGGER_DESCRIPTION',
  SwaggerVersion = 'SWAGGER_VERSION',
  SwaggerPath = 'SWAGGER_PATH',
  DbType = 'DB_TYPE',
  DbHost = 'DB_HOST',
  DbPort = 'DB_PORT',
  DbUser = 'DB_USER',
  DbPassword = 'DB_PASSWORD',
  DbDatabase = 'DB_DATABASE',
  DbSync = 'DB_SYNC',
  DbMigration = 'DB_MIGRATION',
  DbLog = 'DB_LOG',
  DbSchema = 'DB_SCHEMA',
  AuthPasswordMinLength = 'AUTH_PASSWORD_MIN_LENGTH',
  AuthPasswordMaxLength = 'AUTH_PASSWORD_MAX_LENGTH',
  AuthPasswordArgon2MemoryCost = 'AUTH_PASSWORD_ARGON2_MEMORY_COST',
  AuthPasswordArgon2TimeCost = 'AUTH_PASSWORD_ARGON2_TIME_COST',
  AuthPasswordArgon2Parallelism = 'AUTH_PASSWORD_ARGON2_PARALLELISM',
  AuthPasswordMaxAttempts = 'AUTH_PASSWORD_MAX_ATTEMPTS',
  AuthPasswordLockoutSeconds = 'AUTH_PASSWORD_LOCKOUT_SECONDS',
  AuthAccessTokenTtlSeconds = 'AUTH_ACCESS_TOKEN_TTL_SECONDS',
  AuthRefreshTokenTtlSeconds = 'AUTH_REFRESH_TOKEN_TTL_SECONDS',
  AuthSessionAbsoluteTtlSeconds = 'AUTH_SESSION_ABSOLUTE_TTL_SECONDS',
  AuthRefreshTokenPepper = 'AUTH_REFRESH_TOKEN_PEPPER',
  AuthJwtActiveKid = 'AUTH_JWT_ACTIVE_KID',
  AuthJwtPrivateKeyBase64 = 'AUTH_JWT_PRIVATE_KEY_BASE64',
  AuthJwtPublicKeysJson = 'AUTH_JWT_PUBLIC_KEYS_JSON',
  AuthTestRefreshFailurePoint = 'AUTH_TEST_REFRESH_FAILURE_POINT',
}