// src/common/config/env.service.ts
import { Injectable } from '@nestjs/common';
import { ConfigService } from '@nestjs/config';
import { AppMode } from './data/enum/app-mode.enum';
import { ConfigKey } from './data/enum/config-key.enum';
import { DatabaseType } from './data/enum/database-type.enum';
import { LogLevel } from './data/enum/log-level.enum';
import { ValidatedEnvironment } from './environment/environment.validation';

@Injectable()
export class EnvService {
  constructor(
    private readonly configService: ConfigService<ValidatedEnvironment, true>,
  ) {}

  get appMode(): AppMode {
    return this.get(ConfigKey.NodeEnv);
  }

  get appName(): string {
    return this.get(ConfigKey.AppName);
  }

  get appPort(): number {
    return this.get(ConfigKey.AppPort);
  }

  get appBaseUrl(): string {
    return this.get(ConfigKey.AppBaseUrl);
  }

  get corsOrigins(): string[] {
    return this.get(ConfigKey.AppCorsOrigin);
  }

  get trustProxy(): boolean {
    return this.get(ConfigKey.AppTrustProxy);
  }

  get httpPayloadErrorStatusCode(): number {
    return this.get(ConfigKey.AppHttpPayloadErrorCode);
  }

  get logLevel(): LogLevel {
    return this.get(ConfigKey.LogLevel);
  }

  get swaggerEnabled(): boolean {
    return this.get(ConfigKey.SwaggerEnabled);
  }

  get swaggerTitle(): string {
    return this.get(ConfigKey.SwaggerTitle);
  }

  get swaggerDescription(): string {
    return this.get(ConfigKey.SwaggerDescription);
  }

  get swaggerVersion(): string {
    return this.get(ConfigKey.SwaggerVersion);
  }

  get swaggerPath(): string {
    return this.get(ConfigKey.SwaggerPath);
  }

  get databaseType(): DatabaseType {
    return this.get(ConfigKey.DbType);
  }

  get databaseHost(): string {
    return this.get(ConfigKey.DbHost);
  }

  get databasePort(): number {
    return this.get(ConfigKey.DbPort);
  }

  get databaseUser(): string {
    return this.get(ConfigKey.DbUser);
  }

  get databasePassword(): string {
    return this.get(ConfigKey.DbPassword);
  }

  get databaseName(): string {
    return this.get(ConfigKey.DbDatabase);
  }

  get databaseSynchronize(): boolean {
    return this.get(ConfigKey.DbSync);
  }

  get databaseMigrationsRun(): boolean {
    return this.get(ConfigKey.DbMigration);
  }

  get databaseLogging(): boolean {
    return this.get(ConfigKey.DbLog);
  }

  get databaseSchema(): string {
    return this.get(ConfigKey.DbSchema);
  }

  get passwordMinLength(): number {
    return this.get(ConfigKey.AuthPasswordMinLength);
  }

  get passwordMaxLength(): number {
    return this.get(ConfigKey.AuthPasswordMaxLength);
  }

  get passwordArgon2MemoryCost(): number {
    return this.get(ConfigKey.AuthPasswordArgon2MemoryCost);
  }

  get passwordArgon2TimeCost(): number {
    return this.get(ConfigKey.AuthPasswordArgon2TimeCost);
  }

  get passwordArgon2Parallelism(): number {
    return this.get(ConfigKey.AuthPasswordArgon2Parallelism);
  }

  get passwordMaxAttempts(): number {
    return this.get(ConfigKey.AuthPasswordMaxAttempts);
  }

  get passwordLockoutSeconds(): number {
    return this.get(ConfigKey.AuthPasswordLockoutSeconds);
  }

  get accessTokenTtlSeconds(): number {
    return this.get(ConfigKey.AuthAccessTokenTtlSeconds);
  }

  get refreshTokenTtlSeconds(): number {
    return this.get(ConfigKey.AuthRefreshTokenTtlSeconds);
  }

  get sessionAbsoluteTtlSeconds(): number {
    return this.get(ConfigKey.AuthSessionAbsoluteTtlSeconds);
  }

  get refreshTokenPepper(): string {
    return this.get(ConfigKey.AuthRefreshTokenPepper);
  }

  get jwtActiveKid(): string {
    return this.get(ConfigKey.AuthJwtActiveKid);
  }

  get jwtPrivateKeyBase64(): string {
    return this.get(ConfigKey.AuthJwtPrivateKeyBase64);
  }

  get jwtPublicKeysJson(): string {
    return this.get(ConfigKey.AuthJwtPublicKeysJson);
  }

  get testRefreshFailurePoint(): string {
    return this.get(ConfigKey.AuthTestRefreshFailurePoint);
  }

  get isProduction(): boolean {
    return this.appMode === AppMode.Prod;
  }

  get isTest(): boolean {
    return this.appMode === AppMode.Test;
  }

  get<TConfigKey extends ConfigKey>(
    key: TConfigKey,
  ): ValidatedEnvironment[TConfigKey] {
    return this.configService.getOrThrow(key, { infer: true });
  }
}